Rename xas query operators file
